[Fixed] Maintenance and power for factories
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@4404 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/structure/StructureObject.h
@@ -222,13 +222,13 @@ public:
 
 	void setSurplusMaintenance(int surplus);
 
-	void addMaintenance(int add);
+	void addMaintenance(float add);
 
 	void payMaintenance(int maintenance, CreatureObject* payer, bool cashFirst);
 
 	void setSurplusPower(int surplus);
 
-	void addPower(int add);
+	void addPower(float add);
 
 	bool isDecaying();
 
@@ -400,13 +400,13 @@ public:
 
 	void setSurplusMaintenance(int surplus);
 
-	void addMaintenance(int add);
+	void addMaintenance(float add);
 
 	void payMaintenance(int maintenance, CreatureObject* payer, bool cashFirst);
 
 	void setSurplusPower(int surplus);
 
-	void addPower(int add);
+	void addPower(float add);
 
 	bool isDecaying();
 
@@ -559,13 +559,13 @@ public:
 
 	void setSurplusMaintenance(int surplus);
 
-	void addMaintenance(int add);
+	void addMaintenance(float add);
 
 	void payMaintenance(int maintenance, CreatureObject* payer, bool cashFirst);
 
 	void setSurplusPower(int surplus);
 
-	void addPower(int add);
+	void addPower(float add);
 
 	bool isDecaying();
